Fixed compiler type cast warning

+++ base/vulkanMeshLoader.hpp
@@ -130,7 +130,7 @@ namespace vkMeshLoader
 			vertexInputState = vkTools::initializers::pipelineVertexInputStateCreateInfo();
 			vertexInputState.vertexBindingDescriptionCount = 1;
 			vertexInputState.pVertexBindingDescriptions = &bindingDescription;
-			vertexInputState.vertexAttributeDescriptionCount = attributeDescriptions.size();
+			vertexInputState.vertexAttributeDescriptionCount = static_cast<uint32_t>(attributeDescriptions.size());
 			vertexInputState.pVertexAttributeDescriptions = attributeDescriptions.data();
 		}
